requirements.txt
docopt
rauth
cachecontrol[filecache]
python-magic
